backend/requirements.txt
uvicorn
fastapi
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
python-dotenv
langchain
langchain_openai
dateparser
openai
uagents
fetchai
uagents-core